# api/openapi.yaml
openapi: 3.1.0
info:
  title: Trace HTTP API
  version: 1.0.0
  summary: Local-first JSON API for Trace (opt-in via `trace serve`)
  description: |
    Versioned HTTP/JSON contract for Trace core. Default bind is loopback
    (`127.0.0.1`). Handlers and the browser GUI are adapters only (Law 19) —
    they call the canonical Go library / `.trace/` SQLite; they must not fork
    business logic or create a second source of truth.

    **Non-goals (explicit):**
    - No `POST /rpc` / MCP tools/call mirror for the browser (MCP stays stdio-local)
    - No default unbounded / full-graph dump (including seed export body over HTTP)
    - No multi-tenant hosting, OAuth, or billing routes in this phase
    - No CORS `Access-Control-Allow-Origin: *` (deny by default; same-origin SPA)

    Wave tags: `x-trace-wave: p0 | p1 | defer`. GUI relevance: `x-trace-gui: true`.
  contact:
    name: Trace
  license:
    name: See repository LICENSE
servers:
  - url: http://127.0.0.1:7432
    description: Local opt-in `trace serve` (placeholder port; CLI may choose another)
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/tasks/{task_id}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: Get one task by id
      x-trace-wave: p0
      x-trace-gui: true
      parameters:
        - $ref: "#/components/parameters/TaskIdPath"
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskRow"
        "404":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/reviews/{review_id}:
    get:
      tags: [reviews]
      operationId: getReview
      summary: Get review
      x-trace-wave: p1
      x-trace-gui: true
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewSummary"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Required when `trace serve` binds non-loopback (`--allow-remote`).
        On loopback, authentication may be omitted (loopback-trust). Optional
        loopback token is not required for MVP.
  parameters:
    TaskIdPath:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable code (e.g. VALIDATION_ERROR)
            message:
              type: string
            details:
              description: Optional structured context
              nullable: true
...
    TaskRow:
      type: object
      required: [id, title, work_state]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        work_state:
          type: string
        goal_id:
          type: string
          format: uuid
          nullable: true
        body:
          type: string
          nullable: true
...
    ReviewSummary:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
          nullable: true
        result:
          type: string
          enum: [PASS, FAIL, UNCERTAIN]
          nullable: true
        task_id:
          type: string
          format: uuid
          nullable: true